housekeeping: Editorconfig rules for series.md for indentation
This file has YAML in the frontmatter and Markdown elsewhere. The indentation rules don't matter
much for Markdown. However, TAB characters are not allowed in YAML and this causes an error for
Jekyll when the file is built:

	YAML Exception reading /home/siddharth/personal/blog/series.md: (<unknown>): could not find
	expected ':' while scanning a simple key at line 58 column 5

To avoid this error, this patch defines the indentation rule to start using a space and to use 2 as
the size of an indent. This will allow the usage of > and < inside the editor without any risk of
introducing a TAB due to arbitrary editor settings.